Switch build system to CMake.
Run `make` to build the project. This will use `cmake` to build the
project inside `./build/debug`. The resulting `iqdb` binary will be at
`./build/debug/iqdb`.

Run `make release` to build the project with optimizations turned on.
The project will be built in `./build/release` and the binary will be at
`./build/release/iqdb`.