Remove image type detection code.
Remove most code for detecting the image's filetype, and for parsing the
image's width and height. Now we just fail if a non-JPEG file is
uploaded. This is okay for us because Danbooru only sends JPEGs to IQDB
anyway. Trying to deal with detecting, parsing, and resizing N different
image formats is unnecessary.