Remove image widths and heights.
Remove all code for dealing with image width and heights. Previously
IQDB stored the image width and height as a couple of extra pieces of
metadata attached to posts. This wasn't useful to us, because images are
always downscaled to 128x128 by IQDB, and that was the value saved as
the width/height. Also because we didn't actually use it for anything,
if we needed the width or height we got it from the main Danbooru
database.

There was also some weird stuff with the width and height fields being
overloaded to store masks and flags in certain cases. These weren't used
either.

The width and height fields are still in the image signature data
structure, so the on-disk database format doesn't change, but they're
unused.