Remove delta queue data structure.
Remove the delta queue data structure. This was a vector-like data
structure used to store compressed arrays of integers, to reduce memory
usage of IQDB's image index.

IQDB's main data structure is a set of buckets, where each bucket holds
an array of image ids. Delta queues were used to reduce memory usage of
these arrays. However, the code added a lot of complexity, and the
memory savings weren't that great.

Memory usage is now roughly 480 bytes per image, or 480MB per million
images.

The general technique that was used here is called integer compression.
There are other libraries that do this and that can probably do a better
job:

* https://www.codeproject.com/Tips/1080308/An-Introduction-to-Integer-Compression
* https://github.com/lemire/FastPFor