Fix random 'out of memory' errors and crashes.
Fix requests randomly failing with 'out of memory' exceptions,
eventually leading to IQDB hanging or crashing completely.

These exceptions were thrown by SQLite when executing a `SELECT * FROM
images WHERE post_id = ?` statement during the IQDB query. Despite the
error message, they were really caused by making database calls when the
connection was closed.

The root cause was that the SQLite ORM library isn't thread safe. The
fix is to wrap SQL calls in a mutex. This is less than ideal from a
performance perspective, but ultimately fine since the main bottleneck
is in the IQDB query algorithm, not in the SELECT call.